Product Schema & Routes
I have created the basic files for the mongoose models. The product
model has been worked on more and a route has been created, linked to
the server.

I created a tests directory to store all the .http files that test the
api calls for the backend server and check it's response using the
"REST Client" extension in VSCode. In later commits, I will add this to
the .gitignore file.