Redux implementation
In this commit I implemented Redux to allow for the user to sign in and
save the user's information to be accessed in all components in the
website. Additionally, Redux allows to store an array of items that
represent the items inside the user's shopping cart.

P.S I have added an additional property called "isVerified" to the
User.js schema to verify the user's email after registration.