Backend routes (products.js & manufacturers.js)
I have defined most of the products api calls to the backend that mostly
will be used for displaying the information at the home page of the
website.

I created an directory containing helper functions called "functions",
with a Dates.js files. This file contains helper functions considering
Date objects in Javascript.